+++ apps/cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 - feat: Allow overriding client environment with Dart define
 - fix: Searching parent directories for a `pubspec.yaml` file always halts now
 - fix: Only cross-compile with non-Flutter Dart SDKs
+- chore: Use new Flutter version file
 - chore: Update dependencies
 
 ## 1.0.13